Update dependencies and refactor ConfigContext for improved configuration loading
- Updated @tailwindcss/vite and related packages to version 4.1.12 for enhanced features and bug fixes.
- Refactored ConfigContext to streamline configuration loading by utilizing a separate config loader and context management.
- Removed redundant code related to default configuration handling and caching.
