+++ next/src/utils/useConfig.ts
@@ -1,0 +1,12 @@
+import { createContext, useContext } from "react";
+import type { Config } from "./configLoader";
+
+export const ConfigContext = createContext<Config | undefined>(undefined);
+
+export const useConfig = (): Config => {
+  const context = useContext(ConfigContext);
+  if (!context) {
+    throw new Error("useConfig must be used within a ConfigProvider");
+  }
+  return context;
+};
